+++ Assets/Scripts/HP.cs
@@ -5,7 +5,7 @@ public class HP : MonoBehaviour
 {
     [SerializeField] private int Max_HP = 0;
     [SerializeField] private Image HealthBar = null;
-    [SerializeField] private int Current_HP = 0;
+    [SerializeField] private float Current_HP = 0;
 
     private void Start() {
         if (Max_HP > 0)
@@ -22,7 +22,7 @@ public class HP : MonoBehaviour
         HealthBar.fillAmount = (float)Current_HP/Max_HP;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         Current_HP = Current_HP - damage;
         if (Current_HP <= 0)
